Changed toString() method in class dog
Changed the return string of overriding method toString in class dog to display the type of the context object and not the breed. The original one repeated the breed of the animal, because the parent method returned all of the properties of the current context object. But this context also returned the breed since the parent toString was called with the context of the child object in the overriding method.

Additionaly, the calling of toString() at the end is changed to be explicit.

The modifications make the code and the example conveyed easier to understand.